requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.4
anthropic==0.28.0
pgvector==0.3.2
pydantic==2.7.1
pydantic-settings==2.3.0
